Remove unused function and file
